# apps/interlabs-crm-demo/backend/openapi.yaml
openapi: 3.0.3
info:
  title: Interlab Portal API
  version: 1.0.0
  description: |
    Internal Enterprise Dashboard — CRM + ERP + Realtime Operations Hub for
    PT. Interlab Sentra Solutions Indonesia.

    Covers: Auth, Users & Roles, PO lifecycle (11 stages across 5 divisions),
    Sales, Admin & Log, Finance, Technical, HRGA/Legal, Tax & Insurance,
    Notifications, Chat, Email Templates, Files.

    ## Response envelope
    Every response uses a consistent envelope:
      success: `{ "success": true,  "data": ..., "meta"?: {...} }`
      error:   `{ "success": false, "error": "message", "code"?: "..." }`

    ## RBAC
    Every protected endpoint declares `x-rbac` with the feature_key,
    capability, allowed role_keys, and row-level scope. The RBAC middleware
    enforces `(feature, capability)` against `role_permissions` on every
    request; `scope` is applied to queries via the sameRoleScope /
    row-level-security layer.

    ## PO lifecycle side-effects
    Endpoints marked `x-po-side-effects` MUST, in a single transaction:
      1. Append to `purchase_order_status_history`
      2. Append to `purchase_order_tracking_events`
      3. Update `purchase_orders.current_status`
      4. Emit the declared notification templates (if enabled)
  contact:
    name: Interlab Portal Engineering
    email: zakyindrasatriaputra@gmail.com
servers:
  - url: https://api.interlab-portal.com
    description: Production
  - url: http://localhost:3000
...
security:
  - bearerAuth: []
paths:
...
  /api/users:
    get:
      tags: [Users]
      summary: List users
      x-rbac:
        feature: roles_management
        capability: view_own
        allowed_roles: [superadmin, ceo, sales, admin_log, finance, technical, hrga, tax_insurance]
        scope: same-role-or-global
        note: Superadmin/CEO see all; other roles see only users whose role matches their managed_role_scope
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/LimitParam'
        - $ref: '#/components/parameters/SearchParam'
        - { in: query, name: role, schema: { $ref: '#/components/schemas/RoleKey' } }
        - { in: query, name: account_status, schema: { type: string, enum: [active, inactive, suspended] } }
      responses:
        '200':
          description: Paginated users
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessEnvelope'
                  - properties:
                      data:
                        type: array
                        items: { $ref: '#/components/schemas/User' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
    post:
      tags: [Users]
      summary: Create user
      x-rbac:
        feature: roles_management
        capability: create
        allowed_roles: [superadmin, ceo, sales, admin_log, finance, technical, hrga, tax_insurance]
        scope: same-role-or-global
        note: Same-role managers may only create users with matching role_key
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/UserCreateRequest' }
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessEnvelope'
                  - properties:
                      data: { $ref: '#/components/schemas/User' }
        '400': { $ref: '#/components/responses/ValidationError' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409': { $ref: '#/components/responses/Conflict' }
  /api/users/{id}:
    parameters:
      - $ref: '#/components/parameters/UUIDPathId'
    get:
      tags: [Users]
      summary: Get user by id
      x-rbac:
        feature: roles_management
        capability: view_own
        allowed_roles: [superadmin, ceo, sales, admin_log, finance, technical, hrga, tax_insurance]
        scope: same-role-or-global
      responses:
        '200':
          description: User
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessEnvelope'
                  - properties:
                      data: { $ref: '#/components/schemas/User' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
    put:
      tags: [Users]
      summary: Update user
      x-rbac:
        feature: roles_management
        capability: edit
        allowed_roles: [superadmin, ceo, sales, admin_log, finance, technical, hrga, tax_insurance]
        scope: same-role-or-global
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/UserUpdateRequest' }
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessEnvelope'
                  - properties:
                      data: { $ref: '#/components/schemas/User' }
        '400': { $ref: '#/components/responses/ValidationError' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
    delete:
      tags: [Users]
      summary: Soft-delete user
      x-rbac:
        feature: roles_management
        capability: delete
        allowed_roles: [superadmin, ceo]
        scope: global
      responses:
        '204': { description: Deleted }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }

  # ====================================================================
  # Roles & Permissions
  # ====================================================================
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT access token (1h expiry). Obtained from `POST /api/auth/login`.
        Refresh with `POST /api/auth/refresh` (refresh token 7d, or 30d with remember_me).
  parameters:
    PageParam:
      in: query
      name: page
      required: false
      schema: { type: integer, minimum: 1, default: 1 }
    LimitParam:
      in: query
      name: limit
      required: false
      schema: { type: integer, enum: [25, 50, 75, 100], default: 25 }
    SearchParam:
      in: query
      name: search
      required: false
      schema: { type: string }
...
    UUIDPathId:
      in: path
      name: id
      required: true
      schema: { type: string, format: uuid }
  responses:
    Unauthorized:
      description: Missing or invalid JWT
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
    Forbidden:
      description: RBAC denied (role lacks feature.capability, or out-of-scope record)
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
    NotFound:
      description: Resource not found (or soft-deleted)
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
    ValidationError:
      description: Request body / params failed validation
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
    Conflict:
      description: Conflict with current state (e.g. duplicate unique key, invalid stage transition)
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
...
  schemas:
...
    SuccessEnvelope:
      type: object
      required: [success, data]
      properties:
        success: { type: boolean, enum: [true] }
        data: {}
        meta: { $ref: '#/components/schemas/PaginationMeta' }
    ErrorEnvelope:
      type: object
      required: [success, error]
      properties:
        success: { type: boolean, enum: [false] }
        error: { type: string }
        code: { type: string }
        field_errors:
          type: array
          items:
            type: object
            properties:
              field: { type: string }
              message: { type: string }
    PaginationMeta:
      type: object
      properties:
        page: { type: integer }
        limit: { type: integer }
        total: { type: integer }
        total_pages: { type: integer }

    # ------------------------------------------------------------------
    # Enums
    # ------------------------------------------------------------------
    RoleKey:
      type: string
      enum: [superadmin, ceo, sales, admin_log, finance, technical, hrga, tax_insurance]
...
    User:
      type: object
      required: [id, email, role, display_name, account_status]
      properties:
        id: { type: string, format: uuid }
        email: { type: string, format: email }
        role: { $ref: '#/components/schemas/RoleKey' }
        permission_level: { type: string, nullable: true }
        avatar_url: { type: string, nullable: true }
        display_name: { type: string }
        account_status:
          type: string
          enum: [active, inactive, suspended]
        managed_role_scope: { type: string, nullable: true }
        can_manage_same_role: { type: boolean }
        feature_permission_scope: { type: string, nullable: true }
        created_at: { type: string, format: date-time, readOnly: true }
        updated_at: { type: string, format: date-time, readOnly: true }
    UserCreateRequest:
      type: object
      required: [email, password, role, display_name]
      properties:
        email: { type: string, format: email }
        password: { type: string, format: password, minLength: 8 }
        role: { $ref: '#/components/schemas/RoleKey' }
        display_name: { type: string }
        permission_level: { type: string }
        avatar_file_id: { type: string, format: uuid, nullable: true }
        account_status:
          type: string
          enum: [active, inactive, suspended]
          default: active
        managed_role_scope: { type: string }
        can_manage_same_role: { type: boolean, default: false }
        feature_permission_scope: { type: string }
    UserUpdateRequest:
      type: object
      properties:
        email: { type: string, format: email }
        display_name: { type: string }
        permission_level: { type: string }
        avatar_file_id: { type: string, format: uuid, nullable: true }
        account_status:
          type: string
          enum: [active, inactive, suspended]
        password: { type: string, format: password, minLength: 8 }
        managed_role_scope: { type: string }
        can_manage_same_role: { type: boolean }
        feature_permission_scope: { type: string }

    # ------------------------------------------------------------------
    # Role + permissions
    # ------------------------------------------------------------------
...